add linters
