--- src/reducers/index.js
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import greetingReducer from './greetingReducer';
-
-export default combineReducers({
-  greeting: greetingReducer
-});
