--- src/redux/configureStore.js
@@ -1,10 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import messageSlice from './messageSlice';
-
-const store = configureStore({
-  reducer: {
-    message: messageSlice,
-  },
-});
-
-export default store;
